Enable flamegraph generation with pprof
